+++ eiam-20211201/alibabacloud_eiam20211201/models/_create_credential_request.py
@@ -13,10 +13,12 @@ class CreateCredentialRequest(DaraModel):
         credential_identifier: str = None,
         credential_name: str = None,
         credential_scenario_label: str = None,
+        credential_sharing_scope: str = None,
         credential_subject_id: str = None,
         credential_subject_type: str = None,
         credential_type: str = None,
         description: str = None,
+        exclusive_user_id: str = None,
         instance_id: str = None,
     ):
         # 保证请求幂等性。从您的客户端生成一个参数值，确保不同请求间该参数值唯一。ClientToken只支持ASCII字符，且不能超过64个字符。
@@ -37,6 +39,7 @@ class CreateCredentialRequest(DaraModel):
         self.credential_name = credential_name
         # 凭据的使用场景标签。
         self.credential_scenario_label = credential_scenario_label
+        self.credential_sharing_scope = credential_sharing_scope
         # 凭据所属的主体ID。
         self.credential_subject_id = credential_subject_id
         # 凭据所属的主体类型。
@@ -47,6 +50,7 @@ class CreateCredentialRequest(DaraModel):
         self.credential_type = credential_type
         # 描述
         self.description = description
+        self.exclusive_user_id = exclusive_user_id
         # IDaaS EIAM实例的ID。
         # 
         # This parameter is required.
@@ -76,6 +80,9 @@ class CreateCredentialRequest(DaraModel):
         if self.credential_scenario_label is not None:
             result['CredentialScenarioLabel'] = self.credential_scenario_label
 
+        if self.credential_sharing_scope is not None:
+            result['CredentialSharingScope'] = self.credential_sharing_scope
+
         if self.credential_subject_id is not None:
             result['CredentialSubjectId'] = self.credential_subject_id
 
@@ -87,6 +94,9 @@ class CreateCredentialRequest(DaraModel):
 
         if self.description is not None:
             result['Description'] = self.description
+
+        if self.exclusive_user_id is not None:
+            result['ExclusiveUserId'] = self.exclusive_user_id
 
         if self.instance_id is not None:
             result['InstanceId'] = self.instance_id
@@ -111,6 +121,9 @@ class CreateCredentialRequest(DaraModel):
         if m.get('CredentialScenarioLabel') is not None:
             self.credential_scenario_label = m.get('CredentialScenarioLabel')
 
+        if m.get('CredentialSharingScope') is not None:
+            self.credential_sharing_scope = m.get('CredentialSharingScope')
+
         if m.get('CredentialSubjectId') is not None:
             self.credential_subject_id = m.get('CredentialSubjectId')
 
@@ -122,6 +135,9 @@ class CreateCredentialRequest(DaraModel):
 
         if m.get('Description') is not None:
             self.description = m.get('Description')
+
+        if m.get('ExclusiveUserId') is not None:
+            self.exclusive_user_id = m.get('ExclusiveUserId')
 
         if m.get('InstanceId') is not None:
             self.instance_id = m.get('InstanceId')